JWT access token & user login

+++ controllers/contactController.js
@@ -4,7 +4,7 @@ const Contact = require("../models/contactsModel")
 // get all contacts
 //routes GET /api/contacts
 const getContacts = asyncHandler(async (req, res) => {
-  const contacts = await Contact.find();
+  const contacts = await Contact.find({user_id: req.user.id});
   console.log("contacts",contacts)
   res.status(200).json(contacts);
 });
@@ -31,7 +31,8 @@ const createContact = asyncHandler(async (req, res) => {
   const contact = await Contact.create({
     name,
     email,
-    phone
+    phone,
+    user_id: req.user.id,
   });
   console.log("contact created",contact);
   res.status(200).json(contact);
@@ -43,6 +44,11 @@ const updateContact = asyncHandler(async (req, res) => {
   if(!contact){
     res.status(404);
     throw new Error("Contact not found");
+  };
+
+  if (contact.user_id.toString() !== req.user.id){
+    res.status(403);
+    throw new Error("User don't have permission to update other user contacts");
   }
   const updatedContact = await Contact.findByIdAndUpdate(
     req.params.id,
@@ -60,6 +66,10 @@ const deleteContact = asyncHandler(async (req, res) => {
     res.status(404);
     throw new Error("Contact not found");
   };
+  if (contact.user_id.toString() !== req.user.id){
+    res.status(403);
+    throw new Error("User don't have permission to update other user contacts");
+  }
   await Contact.findByIdAndDelete(req.params.id);
   res.status(200).json({ message: "Contact deleted successfully" });
 });
